+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod algorithms;
 pub mod math;
+pub mod object;
 pub mod offline_map;
 pub mod sparse_cost_matrix;