+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod algorithms;
 pub mod math;
 pub mod offline_map;
 pub mod sparse_cost_matrix;
+pub mod large_cost_matrix;