+++ src/lib.rs
@@ -1,5 +1,7 @@
 pub mod algorithms;
 pub mod large_cost_matrix;
+pub mod map;
 pub mod math;
+pub mod object;
 pub mod offline_map;
 pub mod sparse_cost_matrix;