+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Unreleased
   with MongoDB-style IDs
 - Add `map::room_type_for_name` function and `map::RoomType` enum for determining the room type
   of rooms in a default sector layout
+- Add `algorithms::floodfill` module with functions for calculating flood-fills
+- Add `LargeCostMatrix` struct for cost matrices that need more than `u8` sized value data
 
 0.21.1 (2023-05-15)
 ===================